[be] feat: GamerStatus 기본 CRUD+필요 메서드 구현 완료

+++ backend/src/main/java/com/ssafy/be/gamer/repository/GamerStatusRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.be.gamer.repository;
 import com.ssafy.be.gamer.model.domain.GamerStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GamerStatusRepository extends JpaRepository<GamerStatus, Long> {
+public interface GamerStatusRepository extends JpaRepository<GamerStatus, Integer> {
 }
